Uta integration

+++ backend/src/main/kotlin/pl/poznan/put/decision_support/decisionmethods/model/domain/UTA/request/UTARequest.kt
@@ -8,6 +8,6 @@ data class UTARequest(
     val criteriaMinMax: List<String>,
     val criteriaNumberOfBreakPoints: List<Int>,
     val epsilon: Double,
-    val performanceTable: List<List<Int>>,
+    val performanceTable: List<List<Float>>,
     val rownamesPerformanceTable: List<String>
 )
